create useSendMessage hook
